% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{.coerce_to_utf8}
\alias{.coerce_to_utf8}
\title{Coerce to UTF8}
\usage{
.coerce_to_utf8(this_text)
}
\arguments{
\item{this_text}{Character scalar; a piece of text to attempt to coerce.}
}
\value{
The text as UTF8.
}
\description{
Coerce to UTF8
}
\keyword{internal}
