% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{.make_unicode_block_regex}
\alias{.make_unicode_block_regex}
\title{Make Regex for Unicode Blocks}
\usage{
.make_unicode_block_regex(unicode_block_name)
}
\arguments{
\item{unicode_block_name}{The name of the unicode block as it appears in
\href{https://en.wikipedia.org/wiki/Unicode_block#List_of_blocks}{the
Wikipedia list of Unicode blocks}.}
}
\value{
A regex wildcard box in square brackets.
}
\description{
Make Regex for Unicode Blocks
}
\keyword{internal}
