% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{.space_regex_selector}
\alias{.space_regex_selector}
\title{Space Text by a Regex Selector}
\usage{
.space_regex_selector(text, regex_selector)
}
\arguments{
\item{text}{Character; text to space.}

\item{regex_selector}{A regular expression that selects the blocks of text
you want to space.}
}
\value{
A character vector the same length as text, with spaces around the
specified blocks.
}
\description{
Space Text by a Regex Selector
}
\keyword{internal}
