% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecemaker-package.R
\docType{package}
\name{piecemaker-package}
\alias{piecemaker}
\alias{piecemaker-package}
\title{piecemaker: Tools for Preparing Text for Tokenizers}
\description{
Tokenizers break text into pieces that are more usable by machine learning models. Many tokenizers share some preparation steps. This package provides those shared steps, along with a simple tokenizer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/macmillancontentscience/piecemaker}
  \item \url{https://macmillancontentscience.github.io/piecemaker/}
  \item Report bugs at \url{https://github.com/macmillancontentscience/piecemaker/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jonathan Bratt \email{jonathan.bratt@macmillan.com} (\href{https://orcid.org/0000-0003-2859-0076}{ORCID})
}

Other contributors:
\itemize{
  \item Bedford Freeman & Worth Pub Grp LLC DBA Macmillan Learning [copyright holder]
}

}
\keyword{internal}
