% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_cjk}
\alias{space_cjk}
\title{Add Spaces Around CJK Ideographs}
\usage{
space_cjk(text)
}
\arguments{
\item{text}{A character vector to clean.}
}
\value{
A character vector the same length as the input text, with spaces
added between ideographs.
}
\description{
To tokenize Chinese, Japanese, and Korean (CJK) characters, it's convenient to
add spaces around the characters.
}
\examples{
to_space <- intToUtf8(13312:13320)
to_space
space_cjk(to_space)
}
