% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space_punctuation}
\alias{space_punctuation}
\title{Add Spaces Around Punctuation}
\usage{
space_punctuation(text, hyphens = TRUE, abbreviations = TRUE)
}
\arguments{
\item{text}{A character vector to clean.}

\item{hyphens}{Logical; should hyphens be treated as punctuation?}

\item{abbreviations}{Logical; should ' between letters be treated as
punctuation?}
}
\value{
A character vector the same length as the input text, with spaces
added around punctuation characters.
}
\description{
To keep punctuation during tokenization, it's convenient to add spacing
around punctuation. This function does that, with options to keep certain
types of punctuation together as part of the word.
}
\examples{
to_space <- "This is some 'gosh-darn' $5 text. Isn't it lovely?"
to_space
space_punctuation(to_space)
space_punctuation(to_space, hyphens = FALSE)
space_punctuation(to_space, abbreviations = FALSE)
}
