% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize_space}
\alias{tokenize_space}
\title{Break Text at Spaces}
\usage{
tokenize_space(text)
}
\arguments{
\item{text}{A character vector to clean.}
}
\value{
The text as a list of character vectors (one vector per element of
\code{text}). Each element of each vector is roughly equivalent to a word.
}
\description{
This is an extremely simple tokenizer, breaking only and exactly on the space
character. This tokenizer is intended to work in tandem with
\code{\link{prepare_text}}, so that spaces are cleaned up and inserted as
necessary before the tokenizer runs. This function and
\code{\link{prepare_text}} are combined together in
\code{\link{prepare_and_tokenize}}.
}
\examples{
tokenize_space("This is some text.")
}
